// src/modules/IeltsReadingQuestion/enum/questionType.enum.ts
export enum ReadingQuestionType {
    LIST_SELECTION = 'LIST_SELECTION',
    MATCHING_HEADINGS = 'MATCHING_HEADINGS',
    MATCHING_ENDINGS = 'MATCHING_ENDINGS',
    
    /*
    listOfObjects
    listOfStatements
    Answer [A,B,C,D]
    */
    MATCHING_INFORMATION = 'MATCHING_INFORMATION',

    MULTIPLE_CHOICE = 'MULTIPLE_CHOICE',
    SENTENCE_COMPLETION = 'SENTENCE_COMPLETION',
    SUMMARY_NOTE_COMPLETION = 'SUMMARY_NOTE_COMPLETION',
    TRUE_FALSE_NOT_GIVEN = 'TRUE_FALSE_NOT_GIVEN',
    YES_NO_NOT_GIVEN = 'YES_NO_NOT_GIVEN',
}